+++ features/downloading_credit_memo_on_customer_order_view.feature
@@ -1,6 +1,6 @@
-@customer_browsing_credit_memos
+@customer_credit_memos
 Feature: Downloading credit memo on customer order view
-    In order to have access to all credit memos related to order refunds
+    In order to save generated credit memos for future usage
     As a Customer
     I want to be able to download credit memos on single order view
 
@@ -12,8 +12,7 @@ Feature: Downloading credit memo on customer order view
         And I am a logged in customer
         And I placed an order "#00000666"
         And I bought a single "Angel T-Shirt"
-        And I addressed it to "Lucifer Morningstar", "Seaside Fwy", "90802" "Los Angeles" in the "United States"
-        And for the billing address of "Mazikeen Lilim" in the "Pacific Coast Hwy", "90806" "Los Angeles", "United States"
+        And I addressed it to "Lucifer Morningstar", "Seaside Fwy", "90802" "Los Angeles" in the "United States" with identical billing address
         And I chose "Free" shipping method with "Cash on Delivery" payment
         And this order is already paid
         And 1st "Angel T-Shirt" product from order "#00000666" has already been refunded
@@ -22,4 +21,4 @@ Feature: Downloading credit memo on customer order view
     Scenario: Seeing credit memo on customer order view
         When I view the summary of the order "#00000666"
         And I download the first credit memo
-        Then the pdf file for this credit memo should be downloaded successfully
+        Then a pdf file should be successfully downloaded
